requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
